use uint32 for locations

+++ contentprovider.go
@@ -174,14 +174,14 @@ func (p *contentProvider) fillChunkMatches(ms []*candidateMatch, numContextLines
 		for _, m := range ms {
 			ranges = append(ranges, Range{
 				Start: Location{
-					ByteOffset: int(m.byteOffset),
+					ByteOffset: m.byteOffset,
 					LineNumber: 1,
-					Column:     utf8.RuneCount(fileName[:m.byteOffset]) + 1,
+					Column:     uint32(utf8.RuneCount(fileName[:m.byteOffset]) + 1),
 				},
 				End: Location{
-					ByteOffset: int(m.byteOffset + m.byteMatchSz),
+					ByteOffset: m.byteOffset + m.byteMatchSz,
 					LineNumber: 1,
-					Column:     utf8.RuneCount(fileName[:m.byteOffset+m.byteMatchSz]) + 1,
+					Column:     uint32(utf8.RuneCount(fileName[:m.byteOffset+m.byteMatchSz]) + 1),
 				},
 			})
 		}
@@ -307,14 +307,14 @@ func (p *contentProvider) fillContentChunkMatches(ms []*candidateMatch, numConte
 
 			ranges = append(ranges, Range{
 				Start: Location{
-					ByteOffset: int(startOffset),
-					LineNumber: startLine,
-					Column:     utf8.RuneCount(data[startLineOffset:startOffset]) + 1,
+					ByteOffset: startOffset,
+					LineNumber: uint32(startLine),
+					Column:     uint32(utf8.RuneCount(data[startLineOffset:startOffset]) + 1),
 				},
 				End: Location{
-					ByteOffset: int(endOffset),
-					LineNumber: endLine,
-					Column:     utf8.RuneCount(data[endLineOffset:endOffset]) + 1,
+					ByteOffset: endOffset,
+					LineNumber: uint32(endLine),
+					Column:     uint32(utf8.RuneCount(data[endLineOffset:endOffset]) + 1),
 				},
 				SymbolInfo: si,
 			})
@@ -329,8 +329,8 @@ func (p *contentProvider) fillContentChunkMatches(ms []*candidateMatch, numConte
 		chunkMatches = append(chunkMatches, ChunkMatch{
 			Content: newlines.getLines(data, firstLineNumber, chunk.lastLine+numContextLines+1),
 			ContentStart: Location{
-				ByteOffset: int(firstLineStart),
-				LineNumber: firstLineNumber,
+				ByteOffset: firstLineStart,
+				LineNumber: uint32(firstLineNumber),
 				Column:     1,
 			},
 			FileName: false,
@@ -498,8 +498,8 @@ func (p *contentProvider) chunkMatchScore(secs []DocumentSection, m *ChunkMatch,
 
 	for _, r := range m.Ranges {
 		// calculate the start and end offset relative to the start of the content
-		relStartOffset := r.Start.ByteOffset - m.ContentStart.ByteOffset
-		relEndOffset := r.End.ByteOffset - m.ContentStart.ByteOffset
+		relStartOffset := int(r.Start.ByteOffset - m.ContentStart.ByteOffset)
+		relEndOffset := int(r.End.ByteOffset - m.ContentStart.ByteOffset)
 
 		startBoundary := relStartOffset < len(m.Content) && (relStartOffset == 0 || byteClass(m.Content[relStartOffset-1]) != byteClass(m.Content[relStartOffset]))
 		endBoundary := relEndOffset > 0 && (relEndOffset == len(m.Content) || byteClass(m.Content[relEndOffset-1]) != byteClass(m.Content[relEndOffset]))
